Refactored slider handling in mne_visualizer.
No longer handle slider syncing with other objects in mne_visualizer.
instead of ever NOT returning a slider with the graph to the user, ALWAYS return the slider.
If the user wants to hide the slider, they can via the new API - show_ch_slider / show_time_slider.
If False, we just hide the sliders div.
Now, a slider is only ever an input for its own associated graph, in a callback.