MAINT: Update documentation and logging (#155)
* MAINT: Update documentation and logging

- Nice new landing page
- Added covenience makefile commands
- improved logging of flagged epochs

* FIX, MAINT: Bump iclabel version and ignore mne warning

We already throw an error if the user has mne iclabel version less than .5, so added a pin for minimum version of .5 in the requirements file. MNE is also throwing a future warnign for an API kwarg change in epochs.get_data, where the current default copy=False is changing to copy=True. See https://github.com/mne-tools/mne-python/pull/12121 . I'm a little worried about the additional copies we will incur every single time that we call that method, but I dont have a strong reason to stray from what MNE believes is a bug fix so lets go with it and see how it plays.

* FIX, MAINT: Bumpg GH actions checkout versions

- This should resolve the core issue of the runner not being able to find the most up to date
mne_iclabel wheel.

* FIX, TST: Install torch in doc building CI

Since MNE iclabel 0.5, torch is NOT installed by default.