Test on Windows, Linux, MacOS Intel (#169)
* Test on Windows, Linux, MacOS Intel

* FIX: Make install work on winodows

I think that a character encoding error in the README is causing the error on windows.

* FIX: macOS intel warnings

I think the issues is that torch dropped support for macOS Intel upon release of 2.3. And torch < 2.3 doesnt play well with Numpy 2. So the fix here is to downgrade numpy versions to 1 for macOS Intel only.

* FIX: Dont need to upload code coverage for every OS